requirements.txt
google-cloud-bigquery==3.38.0
google-auth==2.41.1
google-api-core==2.27.0
